+++ pysph/base/utils.py
@@ -287,6 +287,9 @@ def get_particle_array_iisph(constants=None, **props):
 def get_particle_array_rigid_body(constants=None, **props):
     """Return a particle array for a rigid body motion.
 
+    For multiple bodies, add a body_id property starting at index 0 with each
+    index denoting the body to which the particle corresponds to.
+
     Parameters
     ----------
     constants : dict
@@ -303,32 +306,37 @@ def get_particle_array_rigid_body(constants=None, **props):
 
     """
     extra_props = ['au', 'av', 'aw', 'V', 'fx', 'fy', 'fz', 'x0', 'y0', 'z0']
+
+    body_id = props.pop('body_id', None)
+    nb = 1 if body_id is None else numpy.max(body_id) + 1
+
     consts = {'total_mass':0.0,
-              'cm': [0.0, 0.0, 0.0],
+              'num_body': numpy.asarray(nb, dtype=int),
+              'cm': numpy.zeros(3*nb, dtype=float),
 
               # The mi are also used to temporarily reduce mass (1), center of
               # mass (3) and the interia components (6), total force (3), total
               # torque (3).
-              'mi': [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
-                     0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-              'force': [0.0, 0.0, 0.0],
-              'torque': [0.0, 0.0, 0.0],
+              'mi': numpy.zeros(16*nb, dtype=float),
+              'force': numpy.zeros(3*nb, dtype=float),
+              'torque': numpy.zeros(3*nb, dtype=float),
               # velocity, acceleration of CM.
-              'vc': [0.0, 0.0, 0.0],
-              'ac': [0.0, 0.0, 0.0],
-              'vc0': [0.0, 0.0, 0.0],
+              'vc': numpy.zeros(3*nb, dtype=float),
+              'ac': numpy.zeros(3*nb, dtype=float),
+              'vc0': numpy.zeros(3*nb, dtype=float),
               # angular velocity, acceleration of body.
-              'omega': [0.0, 0.0, 0.0],
-              'omega0': [0.0, 0.0, 0.0],
-              'omega_dot': [0.0, 0.0, 0.0]
+              'omega': numpy.zeros(3*nb, dtype=float),
+              'omega0': numpy.zeros(3*nb, dtype=float),
+              'omega_dot': numpy.zeros(3*nb, dtype=float)
               }
     if constants:
         consts.update(constants)
     pa = get_particle_array(constants=consts, additional_props=extra_props,
                             **props)
+    pa.add_property('body_id', type='int', data=body_id)
     pa.set_output_arrays( ['x', 'y', 'z', 'u', 'v', 'w', 'rho', 'h', 'm',
                            'p', 'pid', 'au', 'av', 'aw', 'tag', 'gid', 'V',
-                           'fx', 'fy', 'fz'] )
+                           'fx', 'fy', 'fz', 'body_id'] )
     return pa
 
 def get_particle_array_tvf_fluid(constants=None, **props):