+++ CHANGES.rst
@@ -3,7 +3,7 @@
 
 * Release date: not yet released.
 * Fixed major bug with TVF implementation and add support for 3D simulations.
-
+* Add support to easily handle multiple rigid bodies.
 
 1.0a3
 ------